requirements.txt
arrow==0.15.7
click==7.1.2
DateTime==4.3
future==0.18.2
iso8601==0.1.12
pkg-resources==0.0.0
pyfiglet==0.8.post1
PyMySQL==0.9.3
pypi-xmlrpc==0.0.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
RPi.GPIO==0.7.0
serial==0.0.97
six==1.15.0
smbus==1.1.post2
times==0.7
zope.interface==5.1.0
